IEN-539 | allow data-extract role to access reporting page
