+++ apps/web/src/components/milestone-logs/recruitment/Milestone.tsx
@@ -147,7 +147,7 @@ export const EditMilestone: React.FC<EditMilestoneProps> = props => {
         <img src={deleteIcon.src} alt='delete milestone' />
       </button>
     ) : (
-      <button className='pointer-events-none'>
+      <button className='pointer-events-none' data-cy='delete milestone'>
         <img src={disabledDeleteIcon.src} alt='disabled delete milestone' />
       </button>
     );
