+++ packages/common/src/helper/has-access.ts
@@ -1,10 +1,11 @@
 import { Access } from '../enum';
 import { Role } from '../ro';
 
-export const hasAccess = (roles: Role[], acl: Access[]): boolean => {
-  return acl.every(access => {
+export const hasAccess = (roles: Role[], acl: Access[], and = true): boolean => {
+  const condition = (access: Access) => {
     return roles?.some(role => {
       return role.acl?.some(({ slug }) => slug === access);
     });
-  });
+  };
+  return and ? acl.every(condition) : acl.some(condition);
 };
